requirements.txt
asgiref==3.7.2
attrs==23.1.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
fonttools==4.42.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
matplotlib==3.7.2
mysqlclient==2.2.0
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
